Initial planning for usage
My intention is that this program will be something you can call up when
beginning a task, and just say "I'm starting this task" and then forget
about it for a while. No running processes taking up CPU or problematic
buttons clicking and unclicking when you don't want them too.

When you've worked on a project for a while and want to make sure you
remember what you've done, you'll want to add information to your task.
Such as saying "fixed the problem with the null pointer" "fixed memory
leak from not using an indirection pointer with free" and those types of
things. It should be EASY to do this and not take more than a few simple
commands and the information you want to type in.

Once you're done with a task you should be able to say that it's done
and finish with another simple command. And so you don't forget what
you've done 3 months down the road, it's important to write a small
summary message with the finished task.

Eventually, It would be nice to view the current task going on and see
the relevant information being displayed nicely. Including how long
you've worked on the task.