Inital init function created.
The init function is what's going to kick off every function of tcatch,
this is because we need to create the .tc directory to do anything and
have access to it. Once the init function runs we know for a fact that
we have the proper directory structure to deal with.

A refactor will have to occur to abstract getting the .tc directory
location and making the task directory for the current date. Since we'll
be using the directory each time we create or modify a task.

Also added the valgrind command to check for memory leaks to the readme
file and started a How it works section of the readme. This is not only
for the people who are curious how this program works, but also my
planning area and thinking section.