Abstracted Directory functions
Moved directory handling functions to a tc-directory file and modified
the MakeFile to handle it. tc-init is still the primary file to use as
the main function, but this will probably change once work begins on the
functionality.

Added ignore file to ignore .o files and .out files. I should refactor
the directory structure into a standard library one.