Finished Usage for help dialog. Structured flow.
Finished writing the usage for the current commands. While doing so, had
the idea that perhaps I should have a switch task command to let you
switch between tasks. That would be a good idea. Either that or maybe
the start task command could double as that. I'll probably need a remove
task function as well, but for now I'll focus on getting the basics
working.

Added program flow to the main function that describes (at a very high
level) what each command is going to do. No mention of the current task
record file (I haven't decided what to name it yet). A refactor is
neccesary before continued work. The usage should be pushed off into
it's own files. Also. Wondering if I shouldn't check for the -h flag the
way I'm doing it now.

Currently, all the passed in arguments from the command line are looped
through and checked for - and then for --help or -h (all in O(N) time of
course), but should I only except --help/-h as the first arguments to a
command? Or should -h anywhere in the arguments to the program trigger
the help task? Something to consider.