Refactored Flag checking
Refactored the flag checking into a simple function that can check for a
long or short version in all arguments passed in.

Updated readme with plans for how I'll keep track of the current task.
Intentions are to have a current file pointing at the sha1 of the
current task and its location directory for the start task, then
pointers to any of the intermediate "pause" files at all. And of course
the location of a finished task file or not.

Added TRUE and FALSE defines into tc-init. I think a refactor for
tc-init into its own c and header file and the main program being placed
into its own file is due.