Format defined and task name hashing
Create task structure to store information about an individual task.
Defined new scheme for directory structure for the tc system. Within the
.tc directory there is an indexes folder that contains where the
timeline of all tasks will be. This is your history essentially and will
contain a sequence of all tasks worked on.

Each task file is stored in two files. A .seq file and a .info file. The
sequence file (.seq) is used to stored start,pauses,resumes and finishes
of a task as well as the epoch times neccesary to calculate information
about a task. The information file (.info) is just a simple file that
stores the name of the task (in human readable form) and any
information added by the add-info command/

Began work on the task write function that will do the job of creating a
tasks .seq and .info file and updating the master index files.