Finish task write function
Added file manipulation to create the index file and the
seqeuence/information files. The sequence and information files have
information appended into them. They do rely on the calling function to
instantiate the struct, but that's to be expected. It's not the write
jobs job to populate the task struct, only to write it out.

Added initial default values for a struct into the calling tcatch file
and defined some more constants.