Added read functionality and state -> string func.
Added state to string function to transform the defined constants to a
human readable format. This, along with the read function working
correctly (75% -- no info yet) allows the regular tcatch view command to
work correctly.

Now I just need to have it read information in and allow for the view
command to take a --verbose argument to render the entire informational
tidbits about the task.

A refactor may be in order soon. The tcatch file is getting a bit wordy.