Refactored view code to file
Moved code for showing the 'short' version or long version (depending on
the value of the verbose Flag passed in) to its own file -- that way the
other viewing functions can be placed in there as well (such as doing a
more comprehensive bit of viewing.

Malloc-ing taskInfo's pointer now within the working struct, this means
that that pointer must be freed by calling functions, or calling
functions should not exit willy nilly.

Came up with a scheme for rendering the info file. This is detailed in
two long (but identical) comments within the source code (specifically
super man). Essentially, the info file is never read into a variable to
be manipulated, it is piped from the file out onto the stdio stream. Or
at least, that's the plan.