Added note to self. In regards to --all for view
The view --all command will need to scan the task directory and print
out the information (verbosely or not) to the screen. In order to
properly test this I want to be able to generate a bunch of tasks
without having to manually delete the current task file. (which is
currently the only way you can have more than one task because there's
no switching yet -- but there will be soon! )

So I've added a comment on the --all that informs my future self that I
should come back to it later on when it's easier to generated test data.
Not that a simple script couldn't do the job of generating and removing
the current file, but hey. It's my repository and I'll do what I want.