Fixed verbose flag viewing with no taskname
Before, when executing view --verbose, the taskname would be read as
--verbose instead of the current task's name. This caused the sequence
file to not be found and the view command to behave differently then
from a view with no arguments that WOULD find the current task. So this
is now fixed.

There is no checking on the special case for the current task not
existing and there probably should be. This is not the bst strategy
especially since valgrind now reports a conditional jump which is
troublesome. Handling the view with no task name needs to be flushed out
a bit better and will come with a refactor.