Fixed a few errors, added a validation script
Fixed a few uninitalized value errors (there's still one though) and
added a validate script that simple does a bunch of command through
valgrind so I don't have to type them all out and try to remember which
commands might mess with one thing or another.

There is an irritating issue with this commit's version of view -v (no
task name) in that _find_current_task doesn't go to the sequence file
for information or perform calculations correctly, so the start time is
off as well as all other time calculations. It's rather strange because
it's only the verbose output that is incorrect.