Fixed _find_current_task and flag parsing
Changed the flag parsing to use continue again instead of stopping at a
'-' because breaking causes the task name to be blank, which when
switching to a new task results in the new task being a blank for the
name. Which is not what is desired from the command. The previous
version worked fine with start <task name> -s but not start -s <task
name> and let's face it: Semantically, -s should come before the task
name since you're switching to it.