Fixed switching tasks on tasks that exist already.
Prior to this, switching to a new task always went through the flow of
starting a new task. This meant that the sequence number written out
would be 0. That's problematic for a task that's being resumed with the
--switch/-s flag on the start command. To fix this, we now use a
temporary struct and a existence check against the sequence file for a
task. We're kind of doing our working twice here with reading and
hashing and the like. But until a refactor occurs this is working and
functional and start -s is memory leak free.