Finish Task functionality
Finishing a task is pretty simple. Read the task if you can, write it
out with a finished state and make sure to set the endTime accordingly.
Easy. _tc_task_read takes care of the rest. Displaying the view for the
task to the user is nice too, it's much better than saying something
like "your task has been finished" Of course it's been finished, you can
see that from the view saying Task Complete.

The code could be refactored into the tc-task.c file as finishing a task
belongs there since it's a state change for the task really. It might be
a good idea to allow the finish command to be like view, in that if you
pass no arguments it will finish off the current task.