Current task and finish checking
When finishing a task, if it's the current task. The current file needs
to be removed so a user can begin another task without having to use the
-s flag. It's important to store the current task BEFORE you write the
finished task out becuase the write command creates the current file for
the task. So if you do it after then you'll end up with the two task
names (current and working) being the same and you'll always remove the
current file which will cause finishing task foo and removing the
current file even if you're actual current task was task bar.