Updated Sequence reading algorithm.
The calculated times should be a bit more correct. Although there's
still a bug where finishing a task causes the time worked to set to 0
when viewed. I think it may have to do with the numbers being stored in
time_t or the arithmetics done in the view functions using time_t values
and not ints.