Refactored tcatch for view
Refactored tcatch to use the new view command and cleaned up any memory
leaks/unitialized value errors from valgrind that resulted from it. The
code is a mess right now in tcatch because I'm moving to pushing the
struct handling into the subroutines of the commands. That way there
won't be so many exit(1)'s and free-ing of memory every which way.