Strcpy temporary fix.
Because I haven't refactored task.c enough yet, it's read and find
current functions still do some silly things. This means I have to do
a few silly things for my new refactored code to work with it. The task
reading and writing and task starting/stopping will probably come next
in the refactor.

The silly things I'm refering to are that I have to strcpy the taskname
back and forth. Within the view functions, besides line 30,31,39 it's
very clean on how to get the task name. Once the task reading functions
are refactored those lines will be removed.