Refactored add-info
Refactored add info command into it's own file. Modified tcatch.c to use
the new command and removed old code. Updated makefile to link in new
object file code for the command.

While it's not strictly neccesary to push a single function into it's
own file, I think that if I refactor it later on to allow for things
like add-info --task=<task name> --info=<info> and such it will help
keep the code clean. Also I want tcatch.c to really just be the
delegating overlord file that it's meant to be.