Refactored start command
Refactored start command. The funny thing about this is that I now have
the prototype of main stored in tc-init.h so I can call it in tc-start.
Recursive main calling? By Djikstra! That's not compliant with the
standard!

Quit your whining, it's not explicitely disallowed by the standard and
cc has no problem with it so niether do I. Of course because main is
special I lose my stack information and whatnot so I have to free the
memory before I make the call, but that's ok for the sheer sake of the
pleasure I get in calling main recursively (yes I'm evil).