Refactored Finish command
Moved finish command into it's own fileset. Tcatch is now very simple
and just acts as the delagate to the subroutines. Which is how it should
be.

Still have the problem of a task's running time not being computed/shown
correctly. But now that everything is more localized it should be easier
to track down and fix. This is the steps to produce the problem:

Make a task. View the task, it will be correct
Finish the task. View the task, time spent = 0 (wrong)
Start the task again, View the task, time spent = 0 (wrong)