Fixed view calculation
Fixed the running time calculation for finished processes and all
processes in general. Before this fix it would only work correctly for a
task that had just been started, after it stopped and finished it
wouldn't display the right amount of time worked anymore.

Now the problem remains that when a task is finished, the view
calculation is wrong for viewing the finished task. Watching the read
output, it seems the write from finishing the task adds WAY too much
time to the sequence. Not sure why yet.