Fixed timing issue
Finally figured out what the problem was. The calculations by the
algorithm within read task ARE correct -- I'm not crazy. However, when
starting a new task the time written out wasn't the time of writing but
task.startTime which after the initial set of starts/finishes is still
always the original start time. So the sequence file would always have
the start times set as the original start time and not the intermediate
times. This is what caused the strange doubling or exponential growth of
the calculated time.

Now the only problem is that when finishing the time showed is
incorrect. A recurse might fix that.