Fixed Finished task displayed view
Because the endTime was overwritten with current time and the state was
changed to finished, the call to _tc_displayView would show the task
without taking the final segment of the sequence into account.

By setting the task to be in progress and restoring the old endtime
(which would be the prior to finishing seqTime) back to what they were
before writing the task and then calling the view and rendering it as if
it were still running. The total time worked on the task is showne
correctly.