Added --all support to view command.
Passing --all will now list off all the tasks the user has and display
the information for each. Combinations of --all + --verbose are valid
and will be respected by the code.

Valgrind reports no memory leaks or problems with the current build so
all is stable. This is how things should stay. Code on the develop
branch might be unstable for a short period of time as long as it's
documented and doesn't cause segfaults. But any code released to master
shouldn't have any leaks whatsover. It's not hard to remember to free
your memory. Memory management is not a hard thing to do when you keep
track of your opendir, fopen, and malloc commands. Also, most of the
refactoring done today removed a lot of the sudden exit(1)'s from the
code which allows for much cleaner maintenance of the variables.