Added completion script
Added that auto completes tcatch commands if sourced. Note that tcatch
has to be in your path, so either throw it in the /usr/bin directory and
the completion script into your bash completion directory or simply
source it from bash_profile or something.

It doesn't do as many fancy things as it probably could. We could
detect the current task and prompt the user to switch from it or
whatnot. It'd be an easy:

if [[ -f ~/.tc/current ]] ; then
fi

And throw stuff in. But I'm not sure how to get compgen to return
strings besides the arguments/things it's auto completeing. Yet.