Updated readme for new command
Updated the readme file to document the pause command. Also removed the
bit about a finishing summary and poppinng open a text editor. I'm sure
I could accomplish this pretty easily by either invoking a system
command along with the default text editor which is probably set
somewhere in the environment, but for now it's not in the code so let's
not bring false expectations.

There is one slight bug as a feature problem that I do have. Which is
that the task name info file path is written to the info file. This is
because I'm (supposedly cleverly) storing the path to the information
file in the taskInfo field of the tc_task struct itself. Since the
information is written out each time I write, that ends up being in
there.

To fix this I could probably NOT store the info path in the tc_task
struct, and instead just figure out the info path at write time. Which
would be ok as well. Except that it ends a read to my write task
function. Which isn't a bad thing, just something I wanted to avoid.