Updated readme with idea
I just went and edited a file to remove 6 hours from a task I left on
overnight by accident. It was fairly easy, determine the hash of the
task, go to the sequence file, change the finish punch I had just done
to a pause and subtracted 22500 or so from the epoch time.

It's very easy to do, so I wonder if I couldn't make some type of
interface to be able to do that through tcatch itself, providing a user
with a way to page through the sequence file's 'punches' and insert a
new one or delete an old one and in general fix the sequence if it's
wrong.