Fixed view of finished task
When finishing a task, in order for calculation to be correct the state
needs to be in TC_TASK_STARTED because that's what tips off the
calculation to factor in the end time. Most of the time the end time is
the current time itself so tasks are up to date. To get the end time
into the calculation when you're finishing a task (so the time isn't
only calculated from the last sequence line) the state has to be
started. But then the view will display that the task is in progress
--which isn't really true if you're finishing the task.

So I've added a flag to the display_view command that overrides the task
state at the moment before display. This allows the correct text to
appear when finishing a task and doesn't affect the rest of the program
at all (besides needing to use 3 arguments instead of 2)