Fixed info file name being written to info file
Because the information file name is stored in the taskInfo field of the
tc_task struct, it was being written out to the file whenever a write
occured.

To fix that, checking if the fileHash is part of the taskInfo string is
done before writing out the information. Of course this means that one
cannot add information containing the sha-1 hash of the task name. But
honestly, why would anyone do that? If they really want to put it in
there, open the info file an editor and do it manually.