Fixed view -a task name resolution
Using fscanf was a bad idea, I was only reading the first word of the
first line, and not the full line itself. So instead now I'm reading the
line with fgets which makes more sense since that reads until the
newline.

Of course I'm not too thrilled with line 415 of tc-task.c now, mainly
because I don't enjoy using strlen on something that might be 0 long
because then I'm going to try to index at -1 and then who knows what the
hell will happen.

TECHNICALLY, I shouldn't have to worry because there will always be a
taskname in the info file unless someone goes and deletes it (like an
idiot) but unless they truncate the file entirely then there should at
least be a newline or SOME character in the file. Unless maybe the file
wasn't ever created, but then I'll never have opened the file and won't
be executing the code anyway. I THINK I'm safe, but who knows, it could
crash in some unforeseen way due to cosmic rays for all I know.