Updated validate script (sort of)
Added commented out valgrind command to test the deletion of a task.
It's commented out (that's the sort of part) because I don't like to
run the script and have to enter a y or n becuase I'm lazy. I could use
xargs or something to send the n to the command first, then call it
again with a y to truly test it. But right now I'm not concerned because
I've tested it and it's leak free -- if you don't believe me then you
can do it yourself.