Removed Usage from not found during tc-pause
When there is no current task the usage was being displayed for the
tc-pause command. This has been fixed by simply removing the displayal
of usage. Although, since the pause function is only intended to really
work on the current task, perhaps the call to it with argc > 3 should
just be replaced by a usage display.