char -> int to deal with EOF
When compiling on ARMx86 by @primehunter326 the compiler complained
about:

"scrollback buffer src/tc-delete.c:84:2: error: comparison is always true due to limited range of data type"

Which is caused (I believe) by the use of a char being compared to EOF
which is an int. So I'm updating the char yn to be an int so that
comparison is valid. I don't need a while loop going on for infinity
because someone can't hit newline and only wants to send an EOF.